Version 3.2
I added a feature to make a backup of an encrypted file before writing changes. This helps guard against the situation where you may edit a file and write changes with the wrong password. You can still go back to the previous backup version. The backup file will have the same name as the original file with .bak before the original extension. For example:

    .auth.bfa  -->  .auth.bak.bfa

To turn on backups put the following global definition in your .vimrc file:

    let g:openssl_backup = 1